Day 4, part 2: Solved!
Stumbled a bit on the pid reg ex. Should've written a unit test for that field...
Pretty happy with solution. Nice to get a Result with a list of all the errors.
Would've been even nice to get a Result<Passport, Vec<Error>>, but that required more
Rustiness than I can muster at the moment.