2022: Day 15, part 2: Solved!
As expected, the brute force solution from part 1 is too slow.
Went from O(n2) to O(n) roughly, where n is the nr of rows to search.
Trick was a DFS on each row from start col to end col.
If there is no such path following the sensor coverage, there is a gap,
and we know there is only one such gap.
This could be generalized to the 2-d case and even greater speed up
achieved.